+++ linode/resource_linode_domain.go
@@ -102,7 +102,7 @@ func resourceLinodeDomain() *schema.Resource {
 				Optional:    true,
 			},
 			"tags": &schema.Schema{
-				Type:        schema.TypeList,
+				Type:        schema.TypeSet,
 				Elem:        &schema.Schema{Type: schema.TypeString},
 				Optional:    true,
 				Description: "An array of tags applied to this object. Tags are for organizational purposes only.",
@@ -215,7 +215,7 @@ func resourceLinodeDomainCreate(d *schema.ResourceData, meta interface{}) error 
 	}
 
 	if tagsRaw, tagsOk := d.GetOk("tags"); tagsOk {
-		for _, tag := range tagsRaw.([]interface{}) {
+		for _, tag := range tagsRaw.(*schema.Set).List() {
 			createOpts.Tags = append(createOpts.Tags, tag.(string))
 		}
 	}
@@ -287,7 +287,7 @@ func resourceLinodeDomainUpdate(d *schema.ResourceData, meta interface{}) error 
 
 	if d.HasChange("tags") {
 		tags := []string{}
-		for _, tag := range d.Get("tags").([]interface{}) {
+		for _, tag := range d.Get("tags").(*schema.Set).List() {
 			tags = append(tags, tag.(string))
 		}