+++ linode/resource_linode_volume.go
@@ -68,7 +68,7 @@ func resourceLinodeVolume() *schema.Resource {
 				Computed:    true,
 			},
 			"tags": &schema.Schema{
-				Type:        schema.TypeList,
+				Type:        schema.TypeSet,
 				Elem:        &schema.Schema{Type: schema.TypeString},
 				Optional:    true,
 				Description: "An array of tags applied to this object. Tags are for organizational purposes only.",
@@ -146,7 +146,7 @@ func resourceLinodeVolumeCreate(d *schema.ResourceData, meta interface{}) error 
 	}
 
 	if tagsRaw, tagsOk := d.GetOk("tags"); tagsOk {
-		for _, tag := range tagsRaw.([]interface{}) {
+		for _, tag := range tagsRaw.(*schema.Set).List() {
 			createOpts.Tags = append(createOpts.Tags, tag.(string))
 		}
 	}
@@ -209,7 +209,7 @@ func resourceLinodeVolumeUpdate(d *schema.ResourceData, meta interface{}) error 
 	doUpdate := false
 	if d.HasChange("tags") {
 		tags := []string{}
-		for _, tag := range d.Get("tags").([]interface{}) {
+		for _, tag := range d.Get("tags").(*schema.Set).List() {
 			tags = append(tags, tag.(string))
 		}